requirements.txt
aiohttp==3.8.4
aiohttp-cors==0.7.0
aiosignal==1.3.1
alabaster==0.7.13
alembic==1.11.1
altair==5.0.1
annotated-types==0.5.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-generator==1.10
async-lru==2.0.4
async-timeout==4.0.2
asyncpg==0.28.0
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.0
black==23.3.0
bleach==6.0.0
blinker==1.6.2
blis==0.7.9
bokeh==2.4.3
cachetools==5.3.1
catalogue==2.0.8
certifi==2022.12.7
certipy==0.1.3
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
chromadb==0.3.26
click==8.1.3
clickhouse-connect==0.6.6
colorcet==3.0.1
coloredlogs==15.0.1
comm==0.1.3
confection==0.1.0
contourpy==1.0.7
cryptography==41.0.2
curio==1.6
cycler==0.11.0
cymem==2.0.7
dataclasses-json==0.5.9
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
docrepr==0.2.0
docutils==0.18.1
duckdb==0.8.1
elastic-transport==8.4.0
elasticsearch==8.8.2
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.85.1
fastjsonschema==2.16.3
filelock==3.12.2
filetype==1.2.0
fire==0.5.0
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.6.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-search-results==2.4.2
googleapis-common-protos==1.60.0
graphviz==0.20.1
grpcio==1.53.0
grpcio-tools==1.53.0
h11==0.14.0
hnswlib==0.7.0
holoviews==1.15.4
httpcore==0.17.3
httplib2==0.22.0
httptools==0.6.0
httpx==0.24.1
huggingface-hub==0.16.4
humanfriendly==10.0
hvplot==0.8.3
idna==3.4
imagesize==1.4.1
importlib-metadata==6.8.0
iniconfig==2.0.0
ipykernel==6.22.0
ipyparallel==8.6.1
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.5
isoduration==20.11.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
json5==0.9.14
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-bokeh==3.0.7
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter-telemetry==0.1.0
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterhub==4.0.1
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.6
jupyterlab_server==2.24.0
kiwisolver==1.4.4
langchain==0.0.230
langchainplus-sdk==0.0.20
langcodes==3.3.0
lxml==4.9.3
lxml-stubs==0.4.0
lz4==4.3.2
Mako==1.2.4
manifest-ml==0.0.1
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
murmurhash==1.0.9
mypy==1.4.1
mypy-extensions==1.0.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nbtlib==2.0.4
nest-asyncio==1.5.6
nltk==3.8.1
notebook==6.5.3
notebook_shim==0.2.2
numexpr==2.8.4
numpy==1.25.1
oauthlib==3.2.2
olefile==0.46
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
ordered-set==4.1.0
orjson==3.9.2
outcome==1.2.0
overrides==7.3.1
packaging==23.1
pamela==1.1.0
pandas==2.0.0
pandas-stubs==2.0.2.230605
pandocfilters==1.5.0
panel==0.14.4
param==1.13.0
parso==0.8.3
pathlib==1.0.1
pathspec==0.11.1
pathy==0.10.2
pdf2image==1.16.3
pdflatex==0.1.3
pdfminer.six==20221105
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
plantuml==0.3.0
platformdirs==3.1.1
pluggy==1.2.0
posthog==3.0.1
preshed==3.0.8
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.21.12
psutil==5.9.4
psycopg2==2.9.6
ptyprocess==0.7.0
pulsar-client==3.2.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyct==0.5.0
pydantic==2.0.3
pydantic_core==2.3.0
Pygments==2.14.0
PyLaTeX==1.4.1
pyOpenSSL==23.2.0
pypandoc==1.11
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==6.2.5
pytest-asyncio==0.20.3
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-json-logger==2.0.7
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
pyviz-comms==2.2.1
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.1
QtPy==2.3.0
redis==4.6.0
regex==2023.6.3
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
ruff==0.0.278
safetensors==0.3.1
seaborn==0.12.2
Send2Trash==1.8.0
simpleaichat==0.2.1
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.4
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
Sphinx==6.2.1
sphinx-rtd-theme==1.2.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.49
sqlitedict==2.1.0
srsly==2.4.6
stability-sdk==0.8.4
stack-data==0.6.2
starlette==0.20.4
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.3.0
terminado==0.17.1
testpath==0.6.0
thinc==8.1.10
tiktoken==0.4.0
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
trio==0.22.2
typer==0.9.0
types-beautifulsoup4==4.12.0.5
types-bleach==6.0.0.3
types-html5lib==1.1.11.14
types-openpyxl==3.1.0.13
types-Pillow==10.0.0.1
types-protobuf==4.23.0.1
types-PyMySQL==1.1.0.0
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.10
types-requests==2.31.0.1
types-tqdm==4.65.0.1
types-urllib3==1.26.25.13
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
unstructured==0.8.1
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.22.0
uvloop==0.17.0
vega-datasets==0.9.0
Wand==0.6.11
wasabi==1.1.2
watchfiles==0.19.0
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
Werkzeug==2.3.6
wget==3.2
widgetsnbextension==4.0.6
wikipedia==1.4.0
xlrd==2.0.1
XlsxWriter==3.1.2
xyzservices==2023.2.0
yarl==1.9.2
zipp==3.16.0
zstandard==0.21.0
